Re work shipping-address to follow route structure. Good place to stop.
To remind myself:
order/shipping-address route. Model = order.shippingAddress.
This will outlet order/shipping-address/addresses route. Model  = order.user.addressBook
If the address book is empty we immediately transition to
order/shipping-address/addresses/new route. Model = createRecord('address') so that a user can enter a new address.

The goal is for what ever the user submits (when they click the button at the bottom of the page) is validated and then
applied to the order.

I will need to deal with logic for returning to this page after a shipping address has been chosen.
Essently addresses route and address/new route, check to see if shipping-address route's model (order.shippingAddress)
is present and if so check their addresses and if they are equal, select them. That doesn't make much sense, but muchos
tiredos, I'll figure it out, good luck me.